+++ src/main/java/Model/Patient.java
@@ -21,7 +21,7 @@ public class Patient {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(long ID) {
         this.ID = ID;
     }
 
@@ -90,7 +90,7 @@ public class Patient {
         return phone;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(long phone) {
         this.phone = phone;
     }
